Add a parent class for FPN
