+++ mmdet/models/necks/__init__.py
@@ -1,3 +1,4 @@
+from .lateral_fpn import LateralFPN
 from .fpn import FPN
 
-__all__ = ['FPN']
+__all__ = ['FPN', 'LateralFPN']
