+++ mmdet/models/necks/fpn.py
@@ -1,98 +1,30 @@
 import torch.nn as nn
 import torch.nn.functional as F
-from mmcv.cnn import xavier_init
 
-from ..registry import NECKS
+from .lateral_fpn import LateralFPN
 from ..utils import ConvModule
+from ..registry import NECKS
 
 
 @NECKS.register_module
-class FPN(nn.Module):
+class FPN(LateralFPN):
 
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 num_outs,
-                 start_level=0,
-                 end_level=-1,
-                 add_extra_convs=False,
-                 extra_convs_on_inputs=True,
-                 relu_before_extra_convs=False,
-                 conv_cfg=None,
-                 norm_cfg=None,
-                 activation=None):
-        super(FPN, self).__init__()
-        assert isinstance(in_channels, list)
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.num_ins = len(in_channels)
-        self.num_outs = num_outs
-        self.activation = activation
-        self.relu_before_extra_convs = relu_before_extra_convs
-
-        if end_level == -1:
-            self.backbone_end_level = self.num_ins
-            assert num_outs >= self.num_ins - start_level
-        else:
-            # if end_level < inputs, no extra level is allowed
-            self.backbone_end_level = end_level
-            assert end_level <= len(in_channels)
-            assert num_outs == end_level - start_level
-        self.start_level = start_level
-        self.end_level = end_level
-        self.add_extra_convs = add_extra_convs
-        self.extra_convs_on_inputs = extra_convs_on_inputs
-
-        self.lateral_convs = nn.ModuleList()
+    def __init__(self, *args, **kwargs):
+        super(FPN, self).__init__(*args, **kwargs)
         self.fpn_convs = nn.ModuleList()
 
         for i in range(self.start_level, self.backbone_end_level):
-            l_conv = ConvModule(
-                in_channels[i],
-                out_channels,
-                1,
-                conv_cfg=conv_cfg,
-                norm_cfg=norm_cfg,
-                activation=self.activation,
-                inplace=False)
             fpn_conv = ConvModule(
-                out_channels,
-                out_channels,
+                self.out_channels,
+                self.out_channels,
                 3,
                 padding=1,
-                conv_cfg=conv_cfg,
-                norm_cfg=norm_cfg,
+                conv_cfg=self.conv_cfg,
+                norm_cfg=self.norm_cfg,
                 activation=self.activation,
                 inplace=False)
 
-            self.lateral_convs.append(l_conv)
             self.fpn_convs.append(fpn_conv)
-
-        # add extra conv layers (e.g., RetinaNet)
-        extra_levels = num_outs - self.backbone_end_level + self.start_level
-        if add_extra_convs and extra_levels >= 1:
-            for i in range(extra_levels):
-                if i == 0 and self.extra_convs_on_inputs:
-                    in_channels = self.in_channels[self.backbone_end_level - 1]
-                else:
-                    in_channels = out_channels
-                extra_fpn_conv = ConvModule(
-                    in_channels,
-                    out_channels,
-                    3,
-                    stride=2,
-                    padding=1,
-                    conv_cfg=conv_cfg,
-                    norm_cfg=norm_cfg,
-                    activation=self.activation,
-                    inplace=False)
-                self.fpn_convs.append(extra_fpn_conv)
-
-    # default init_weights for conv(msra) and norm in ConvModule
-    def init_weights(self):
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                xavier_init(m, distribution='uniform')
 
     def forward(self, inputs):
         assert len(inputs) == len(self.in_channels)
@@ -125,12 +57,15 @@ class FPN(nn.Module):
             else:
                 if self.extra_convs_on_inputs:
                     orig = inputs[self.backbone_end_level - 1]
-                    outs.append(self.fpn_convs[used_backbone_levels](orig))
+                    outs.append(self.extra_convs[0](orig))
                 else:
-                    outs.append(self.fpn_convs[used_backbone_levels](outs[-1]))
+                    outs.append(self.extra_convs[0](outs[-1]))
                 for i in range(used_backbone_levels + 1, self.num_outs):
                     if self.relu_before_extra_convs:
-                        outs.append(self.fpn_convs[i](F.relu(outs[-1])))
+                        outs.append(self.extra_convs[i - used_backbone_levels](
+                            F.relu(outs[-1])))
                     else:
-                        outs.append(self.fpn_convs[i](outs[-1]))
+                        outs.append(self.extra_convs[i - used_backbone_levels](
+                            outs[-1]))
+
         return tuple(outs)
